+++ src/main/java/com/shinhanDS5gi/memento/domain/Mentos.java
@@ -1,6 +1,7 @@
 package com.shinhanDS5gi.memento.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.shinhanDS5gi.memento.domain.Reservation.Reservation;
 import com.shinhanDS5gi.memento.domain.base.BaseStatus;
 import com.shinhanDS5gi.memento.domain.base.BaseTime;
 import com.shinhanDS5gi.memento.domain.member.Member;
