+++ src/main/java/com/shinhanDS5gi/memento/domain/member/Member.java
@@ -2,6 +2,7 @@ package com.shinhanDS5gi.memento.domain.member;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.shinhanDS5gi.memento.domain.*;
+import com.shinhanDS5gi.memento.domain.Reservation.Reservation;
 import com.shinhanDS5gi.memento.domain.base.BaseStatus;
 import com.shinhanDS5gi.memento.domain.base.BaseTime;
 import com.shinhanDS5gi.memento.domain.payment.Payment;
