+++ src/main/java/com/shinhanDS5gi/memento/repository/ReservationRepository.java
@@ -1,6 +1,6 @@
 package com.shinhanDS5gi.memento.repository;
 
-import com.shinhanDS5gi.memento.domain.Reservation;
+import com.shinhanDS5gi.memento.domain.Reservation.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
