+++ src/main/java/com/shinhanDS5gi/memento/service/MyPageServiceImpl.java
@@ -1,7 +1,7 @@
 package com.shinhanDS5gi.memento.service;
 
 import com.shinhanDS5gi.memento.common.exception.MemberException;
-import com.shinhanDS5gi.memento.domain.Reservation;
+import com.shinhanDS5gi.memento.domain.Reservation.Reservation;
 import com.shinhanDS5gi.memento.domain.base.BaseStatus;
 import com.shinhanDS5gi.memento.domain.member.Member;
 import com.shinhanDS5gi.memento.domain.member.MemberType;
